requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
pillow==10.0.0
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
psycopg2-binary==2.9.7
redis==5.0.0
celery==5.3.0
django-extensions==3.2.3
coverage==7.3.0

